Rename files for hover-trigger
