+++ public/js/component/hover_trigger.js
