+++ public/scss/style.scss
@@ -12,7 +12,7 @@
 @import 'container';
 @import 'form';
 @import 'header';
-@import 'hover_popup';
+@import 'hover_trigger';
 @import 'image';
 @import 'list';
 @import 'nav';
